Added VisibilityIndex tool
Added VisibilityIndex tool

+++ src/tools/image_analysis/edge_preserving_mean_filter.rs
@@ -156,9 +156,9 @@ impl WhiteboxTool for EdgePreservingMeanFilter {
                 };
             } else if flag_val == "-filter" || flag_val == "-filter" {
                 filter_size = if keyval {
-                    vec[1].to_string().parse::<usize>().unwrap()
+                    vec[1].to_string().parse::<f64>().unwrap() as usize
                 } else {
-                    args[i+1].to_string().parse::<usize>().unwrap()
+                    args[i+1].to_string().parse::<f64>().unwrap() as usize
                 };
             } else if flag_val == "-threshold" {
                 threshold = if keyval {
